~ restart reliably by delaying
Delay restarts by nodemon by 200ms. A delay of 100ms was too short.

Attempt to use forever in combination with `nodemon --exitcrash` did not work. Sometimes the setup would still end up with a situation where the port was still in use. (EADDRINUSE)

Another alternative would be to try node-supervisor.